tst: consolidate Tauri boundary contracts
